Split templating into a cogt part
