+++ pipelex/cogt/templating/template_blueprint.py
@@ -2,9 +2,9 @@ from typing import Any
 
 from pydantic import BaseModel, Field, model_validator
 
+from pipelex.cogt.templating.template_category import TemplateCategory
+from pipelex.cogt.templating.templating_style import TemplatingStyle
 from pipelex.tools.templating.jinja2_parsing import check_jinja2_parsing
-from pipelex.tools.templating.template_category import TemplateCategory
-from pipelex.tools.templating.templating_style import TemplatingStyle
 
 
 class TemplateBlueprint(BaseModel):
