+++ pipelex/core/stuffs/stuff_artefact.py
@@ -4,10 +4,10 @@ from jinja2.runtime import Context
 from pydantic import RootModel
 from typing_extensions import override
 
+from pipelex.cogt.templating.templating_style import TextFormat
 from pipelex.core.stuffs.stuff_content import StuffContent
 from pipelex.exceptions import StuffArtefactError
 from pipelex.tools.templating.jinja2_models import Jinja2ContextKey, Jinja2TaggableAbstract
-from pipelex.tools.templating.templating_style import TextFormat
 
 
 class StuffArtefact(RootModel[dict[str, Any]], Jinja2TaggableAbstract):
