+++ pipelex/tools/templating/jinja2_parsing.py
@@ -1,8 +1,8 @@
 import jinja2
 
+from pipelex.cogt.templating.template_category import TemplateCategory
 from pipelex.tools.templating.jinja2_environment import make_jinja2_env_without_loader
-from pipelex.tools.templating.jinja2_errors import TemplateSyntaxError
-from pipelex.tools.templating.template_category import TemplateCategory
+from pipelex.tools.templating.jinja2_errors import Jinja2TemplateSyntaxError
 
 
 def check_jinja2_parsing(
@@ -14,4 +14,4 @@ def check_jinja2_parsing(
         jinja2_env.parse(jinja2_template_source)
     except jinja2.exceptions.TemplateSyntaxError as exc:
         msg = f"Could not parse Jinja2 template because of: {exc}. Template source:\n{jinja2_template_source}"
-        raise TemplateSyntaxError(msg) from exc
+        raise Jinja2TemplateSyntaxError(msg) from exc
