+++ pipelex/tools/templating/jinja2_rendering.py
@@ -8,17 +8,17 @@ from jinja2.exceptions import (
 )
 
 from pipelex import log
+from pipelex.cogt.templating.template_category import TemplateCategory
+from pipelex.cogt.templating.template_preprocessor import preprocess_template
+from pipelex.cogt.templating.templating_style import TemplatingStyle
 from pipelex.tools.templating.jinja2_environment import make_jinja2_env_without_loader
 from pipelex.tools.templating.jinja2_errors import (
     Jinja2ContextError,
-    Jinja2RenderError,
     Jinja2StuffError,
+    Jinja2TemplateRenderError,
 )
 from pipelex.tools.templating.jinja2_models import Jinja2ContextKey
 from pipelex.tools.templating.jinja2_parsing import check_jinja2_parsing
-from pipelex.tools.templating.template_category import TemplateCategory
-from pipelex.tools.templating.template_preprocessor import preprocess_template
-from pipelex.tools.templating.templating_style import TemplatingStyle
 
 
 def _add_to_templating_context(temlating_context: dict[str, Any], jinja2_context_key: Jinja2ContextKey, value: Any) -> None:
@@ -43,7 +43,7 @@ async def render_jinja2(
         template = jinja2_env.from_string(template_source)
     except TemplateAssertionError as exc:
         msg = f"Jinja2 render error: '{exc}', template_source:\n{template_source}"
-        raise Jinja2RenderError(msg) from exc
+        raise Jinja2TemplateRenderError(msg) from exc
     template_source = preprocess_template(template_source)
     check_jinja2_parsing(
         jinja2_template_source=template_source,
@@ -72,14 +72,14 @@ async def render_jinja2(
         generated_text: str = await template.render_async(**temlating_context)
     except Jinja2StuffError as stuff_error:
         msg = f"Jinja2 render — stuff error: '{stuff_error}', template_source:\n{template_source}"
-        raise Jinja2RenderError(msg) from stuff_error
+        raise Jinja2TemplateRenderError(msg) from stuff_error
     except TemplateSyntaxError as syntax_error:
         msg = f"Jinja2 render — syntax error: '{syntax_error}', template_source:\n{template_source}"
-        raise Jinja2RenderError(msg) from syntax_error
+        raise Jinja2TemplateRenderError(msg) from syntax_error
     except UndefinedError as undef_error:
         msg = f"Jinja2 render — undefined error: '{undef_error}', template_source:\n{template_source}"
-        raise Jinja2RenderError(msg) from undef_error
+        raise Jinja2TemplateRenderError(msg) from undef_error
     except Jinja2ContextError as context_error:
         msg = f"Jinja2 render — context error: '{context_error}', template_source:\n{template_source}"
-        raise Jinja2RenderError(msg) from context_error
+        raise Jinja2TemplateRenderError(msg) from context_error
     return generated_text
