--- pipelex/tools/templating/templates/base_templates.toml
@@ -1,36 +1,0 @@
-[generic_prompts]
-structure_from_preliminary_text_system = "You are a data modeling expert specialized in extracting structure from text."
-# TODO: Should be a proper jinja2
-structure_from_preliminary_text_user = """
-Your job is to extract and structure information from a text.
-Here is the text:
-
-{preliminary_text}
-
-Now generate the JSON in the required format.
-Do not create information that is not in the text.
-"""
-
-output_structure_prompt = """
-
----
-The instance we want to generate will be for the following class:
-{{ class_structure_str }}
-
-Don't bother with JSON formatting, we'll do that as a second step.
-For now, just output markdown with the details of the instance.
-DO NOT create information.
-If some information is not present for an attribute, output the default value or None according to the attribute definition.
-"""
-
-output_structure_prompt_no_preliminary_text = """
----
-The instance we want to generate will be for the following class:
-{{ class_structure_str }}
-
-DO NOT create information.
-If some information is not present for an attribute, output the default value or None according to the attribute definition.
-"""
-
-[test_prompts]
-jinja2_test_template = "I want a {{ place_holder }} t-shirt."
