+++ tests/integration/pipelex/pipes/test_pipe_compose.py
@@ -3,6 +3,9 @@ from typing import TYPE_CHECKING, cast
 import pytest
 
 from pipelex import pretty_print
+from pipelex.cogt.templating.template_blueprint import TemplateBlueprint
+from pipelex.cogt.templating.template_category import TemplateCategory
+from pipelex.cogt.templating.templating_style import TagStyle, TemplatingStyle, TextFormat
 from pipelex.core.concepts.concept_native import NativeConceptCode
 from pipelex.core.memory.working_memory_factory import WorkingMemoryFactory
 from pipelex.hub import get_pipe_router
@@ -11,9 +14,6 @@ from pipelex.pipe_operators.compose.pipe_compose_factory import PipeComposeFacto
 from pipelex.pipe_run.pipe_job_factory import PipeJobFactory
 from pipelex.pipe_run.pipe_run_params import PipeRunMode
 from pipelex.pipe_run.pipe_run_params_factory import PipeRunParamsFactory
-from pipelex.tools.templating.template_blueprint import TemplateBlueprint
-from pipelex.tools.templating.template_category import TemplateCategory
-from pipelex.tools.templating.templating_style import TagStyle, TemplatingStyle, TextFormat
 from tests.cases import JINJA2TestCases
 
 if TYPE_CHECKING:
