+++ tests/unit/pipelex/core/test_data/pipes/operators/compose/pipe_compose.py
@@ -1,8 +1,8 @@
+from pipelex.cogt.templating.template_blueprint import TemplateBlueprint
+from pipelex.cogt.templating.template_category import TemplateCategory
 from pipelex.core.bundles.pipelex_bundle_blueprint import PipelexBundleBlueprint
 from pipelex.core.concepts.concept_native import NativeConceptCode
 from pipelex.pipe_operators.compose.pipe_compose_blueprint import PipeComposeBlueprint
-from pipelex.tools.templating.template_blueprint import TemplateBlueprint
-from pipelex.tools.templating.template_category import TemplateCategory
 
 PIPE_COMPOSE_WITHOUT_CATEGORY = (
     "pipe_compose_without_category",
