+++ tests/unit/pipelex/tools/templating/test_template_preprocessor.py
@@ -1,5 +1,5 @@
 from pipelex import pretty_print
-from pipelex.tools.templating.template_preprocessor import preprocess_template
+from pipelex.cogt.templating.template_preprocessor import preprocess_template
 
 
 class TestTemplatePreprocessor:
